drf_tutorial/requirements.txt
asgiref @ file:///C:/b/abs_ae1n0p_8_w/croot/asgiref_1724072476516/work
Django==4.2
django-filter==25.1
djangorestframework==3.16.1
pillow==11.3.0
sqlparse @ file:///C:/b/abs_68npyz0s6_/croot/sqlparse_1734704476369/work
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
